feat: Support the NZBN bank account number format more explicitly
The IRD document referenced in the README is an implementation of the
NZ styled international banking format - which isn't commonly in use in
New Zealand is pretty useless is a local format where we can make more
assumptions to get a wider range of valid account numbers. Based on the
wikipedia page here: https://en.wikipedia.org/wiki/New_Zealand_bank_account_number
we can be a little more aggressive with some assumptions when we match
the BB-bbbb-AAAAAAA-SSS format outlined.

Add a kiwibank test case number without spacing or hyphens. This number
should be valid, ANZ reckons it is fine during the "pay a friend" screen.
This number fails to validate prior to change, but we have made payments
to this account successful with our real bank.